Typing an invalid date in the report date picker would creash the front end. Fixes #1394.
